Automatically refresh status on changes.
Laebel now subscribes to Docker container events and pushes a Server Side Event to the client, telling it to update the service status. The client uses HTMX to listen for SSE events and refresh the relevant part of the page.

I use CSS to set status icons and details, as this allows me to update the already rendered Mermaid.js graph without modifying it.

I also publish a "reload" SSE event on structural changes to the project (containers added/removed) which requires the client to reload the whole page. Ideally, we should only reload the new parts, but that does not work with Mermaid.js. The server also pushes a "close" event after this, to tell the client to stop expecting new events.